baek

+++ BAEKJOON/10869.cpp
@@ -11,7 +11,7 @@ int main(){
     cout << a + b << endl;
     cout << a - b << endl;
     cout << a * b << endl;
-    cout << (double) a / (double) b << endl;
+    cout << a / b << endl;
     cout << a % b << endl;
 
     return 0;
